+++ server/src/main/java/romanticweapon/server/config/web/WebConfig.java
@@ -1,4 +1,4 @@
-package romanticweapon.server.config;
+package romanticweapon.server.config.web;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
