+++ server/src/main/java/romanticweapon/server/handler/OAuth2SuccessHandler.java
@@ -6,13 +6,12 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 import romanticweapon.server.config.auth.JwtTokenProvider;
-import romanticweapon.server.repository.UserRepository;
+import romanticweapon.server.repository.auth.UserRepository;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.net.URLEncoder;
 
 @Component
 @RequiredArgsConstructor
